Lecture code 1/requirements.txt
Django==5.2.5
pillow==11.3.0
six==1.17.0
djangorestframework==3.16.1
django-filter==25.2
celery==5.5.3
django-celery-beat==2.8.1
redis==6.4.0
